add github actions and mypy

+++ packages/python-client/lmk/process/daemon.py
@@ -8,10 +8,11 @@ import signal
 import socket
 import textwrap
 from functools import wraps
-from typing import Optional, Callable, Any
+from typing import Optional, Callable, Any, cast, Awaitable, AsyncGenerator
 
 from aiohttp import web
 
+from lmk.generated.models.event_response import EventResponse
 from lmk.generated.models.session_response import SessionResponse
 from lmk.generated.models.process_session_state import ProcessSessionState
 from lmk.instance import get_instance
@@ -19,7 +20,14 @@ from lmk.process import exc
 from lmk.process.manager import JobManager
 from lmk.process.models import Job
 from lmk.process.monitor import ProcessMonitor, MonitoredProcess
-from lmk.utils import setup_logging, socket_exists, read_last_lines, shlex_join
+from lmk.utils import (
+    setup_logging,
+    socket_exists,
+    read_last_lines,
+    shlex_join,
+    asyncio_event,
+)
+from lmk.utils.ws import WebSocket
 
 
 LOGGER = logging.getLogger(__name__)
@@ -65,9 +73,9 @@ class ProcessMonitorController:
         self.monitor = monitor
         self.hostname = socket.gethostname()
 
-        self.done_event = asyncio.Event(loop=loop)
-        self.attached_event = asyncio.Event(loop=loop)
-        self.update_event = asyncio.Event(loop=loop)
+        self.done_event = asyncio_event(loop=loop)
+        self.attached_event = asyncio_event(loop=loop)
+        self.update_event = asyncio_event(loop=loop)
         self.session: Optional[SessionResponse] = None
         self.process: Optional[MonitoredProcess] = None
 
@@ -79,7 +87,7 @@ class ProcessMonitorController:
         return False
 
     @route_handler
-    async def _wait_websocket(self, request: web.Request) -> web.Response:
+    async def _wait_websocket(self, request: web.Request) -> web.WebSocketResponse:
         wait_for = request.query.get("wait_for") or "run"
         ws = web.WebSocketResponse()
         await ws.prepare(request)
@@ -89,6 +97,8 @@ class ProcessMonitorController:
             await self.attached_event.wait()
 
             job = await self.manager.get_job(self.job_name)
+            if job is None:
+                raise exc.JobNotFound(self.job_name)
 
             if job.error is not None:
                 await ws.send_json(
@@ -109,6 +119,9 @@ class ProcessMonitorController:
 
         await self.done_event.wait()
         job = await self.manager.get_job(self.job_name)
+
+        if job is None:
+            raise exc.JobNotFound(self.job_name)
 
         if job.error is not None:
             await ws.send_json(
@@ -144,30 +157,41 @@ class ProcessMonitorController:
         return web.json_response({"ok": True})
 
     async def _handle_session_action(self, action: str, body: Optional[Any]) -> None:
-        if action == "sendSignal":
+        if action == "sendSignal" and body is not None:
             await self.send_signal(body["signal"])
             return
         LOGGER.error("Invalid session action: %s, body: %s", action, body)
 
     @contextlib.asynccontextmanager
-    async def _session_ctx(self) -> None:
+    async def _session_ctx(self) -> AsyncGenerator[None, None]:
         instance = get_instance()
         job = await self.manager.get_job(self.job_name)
-        self.session = await instance.create_session(
-            self.job_name,
-            ProcessSessionState(
-                type="process",
-                hostname=self.hostname,
-                command=shlex_join(json.loads(job.command)),
-                pid=job.pid,
-                notifyOn=job.notify_on,
-                notifyChannel=job.channel_id,
+        if job is None:
+            raise exc.JobNotFound(self.job_name)
+
+        self.session = await cast(
+            Awaitable[SessionResponse],
+            instance.create_session(
+                self.job_name,
+                ProcessSessionState(
+                    type="process",
+                    hostname=self.hostname,
+                    command=shlex_join(json.loads(job.command))
+                    if job.command
+                    else "<unknown>",
+                    pid=cast(float, job.pid),
+                    notifyOn=job.notify_on,
+                    notifyChannel=job.channel_id,
+                    exitCode=None,
+                ),
+                async_req=True,
             ),
-            async_req=True,
         )
+
         LOGGER.info("Created session: %s", self.session.session_id)
         await self.manager.update_job(self.job_name, session_id=self.session.session_id)
 
+        ws: WebSocket
         async with instance.session_connect(self.session.session_id, False) as ws:
             LOGGER.debug("Connected to session: %s", self.session.session_id)
 
@@ -242,7 +266,11 @@ class ProcessMonitorController:
                 LOGGER.debug("Waiting for session tasks")
                 await asyncio.gather(updates_task, messages_task)
                 LOGGER.debug("Session tasks are done")
-                await instance.end_session(self.session.session_id, async_req=True)
+                if self.session is not None:
+                    await cast(
+                        Awaitable[None],
+                        instance.end_session(self.session.session_id, async_req=True),
+                    )
 
     async def _run_server(self) -> None:
         socket_path = self.manager.socket_file(self.job_name)
@@ -278,6 +306,9 @@ class ProcessMonitorController:
             job = await self.manager.get_job(self.job_name)
 
             try:
+                if job is None:
+                    raise exc.JobNotFound(self.job_name)
+
                 # Create the output file
                 with open(output_path, "wb+") as f:
                     pass
@@ -288,7 +319,9 @@ class ProcessMonitorController:
                     log_path,
                     log_level,
                 )
-                LOGGER.debug("Attached to %d (%s)", self.process.pid, self.process.command)
+                LOGGER.debug(
+                    "Attached to %d (%s)", self.process.pid, self.process.command
+                )
                 await self.manager.update_job(
                     self.job_name,
                     pid=self.process.pid,
@@ -298,7 +331,10 @@ class ProcessMonitorController:
 
                 await stack.enter_async_context(self._session_ctx())
 
-                LOGGER.debug("Entered session context: %s", self.session.session_id)
+                LOGGER.debug(
+                    "Entered session context: %s",
+                    self.session.session_id if self.session else "<unknown>",
+                )
 
                 exit_code = await self.process.wait()
                 job = await self.manager.end_job(
@@ -318,7 +354,7 @@ class ProcessMonitorController:
             else:
                 LOGGER.info("%d: exited with code %d", job.pid, job.exit_code)
             finally:
-                should_notify = self._should_notify(job)
+                should_notify = self._should_notify(cast(Job, job))
 
             self.done_event.set()
 
@@ -343,9 +379,9 @@ class ProcessMonitorController:
                         ```
                         Process ran on `{self.hostname}`
 
-                        Started: {job.started_at.isoformat()}
+                        Started: {job.started_at.isoformat() if job.started_at else "<unknown>"}
                         
-                        Ended: {job.ended_at.isoformat()}
+                        Ended: {job.ended_at.isoformat() if job.ended_at else "<unknown>"}
                         """
                     ).strip()
 
@@ -354,10 +390,15 @@ class ProcessMonitorController:
                     if logs:
                         message += f"\n\nMost recent logs:\n```\n{logs}\n```"
 
-                    await instance.notify(
-                        message,
-                        notification_channels=job.channel_id and [job.channel_id],
-                        async_req=True,
+                    await cast(
+                        Awaitable[EventResponse],
+                        instance.notify(
+                            message,
+                            notification_channels=(
+                                None if job.channel_id is None else [job.channel_id]
+                            ),
+                            async_req=True,
+                        ),
                     )
                     notify_status = "success"
                 except Exception:
@@ -422,7 +463,9 @@ class ProcessMonitorDaemon(multiprocessing.Process):
         loop = asyncio.new_event_loop()
         asyncio.set_event_loop(loop)
 
-        controller = ProcessMonitorController(self.job_name, self.monitor, manager, loop=loop)
+        controller = ProcessMonitorController(
+            self.job_name, self.monitor, manager, loop=loop
+        )
 
         try:
             log_path = manager.log_file(self.job_name)
@@ -430,8 +473,7 @@ class ProcessMonitorDaemon(multiprocessing.Process):
             with open(log_path, "a+") as log_stream:
                 setup_logging(log_stream=log_stream, level=self.log_level)
 
-                with pid_ctx(pid_file, self.pid):
-
+                with pid_ctx(pid_file, cast(int, self.pid)):
                     try:
                         loop.run_until_complete(
                             controller.run(log_path, self.log_level)
